fix: rename date utils to chat
