+++ src/utils/chat.ts
